# Conflicts:
#	FunnyMoneyClicker/Assets/TextMesh Pro/Examples & Extras/Resources/Fonts & Materials/Bangers SDF.asset

+++ FunnyMoneyClicker/Assets/People/Gabriel/SaveData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 [Serializable]
 public class SaveData
 {
-    public float moneyCount;
+    public double moneyCount;
     public float playTime;
     public int[] upgradeLevels = new int[10]; // Supports up to 10 upgrades
     public int[] powerLevels = new int[10]; // Supports up to 10 upgrades
